digitalwallet/requirements/base.txt
amqp==5.0.6
asgiref==3.4.0
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
django-braces==1.14.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-guardian==2.4.0
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
kombu==5.1.0
prompt-toolkit==3.0.19
pycodestyle==2.7.0
python-dateutil==2.8.2
python-http-client==3.3.2
pytz==2021.1
sendgrid==6.7.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
